requirements.txt
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai==1.50.2
openai-whisper==20240927
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytubefix==7.1.3
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
text-unidecode==1.3
tiktoken==0.7.0
tinycss2==1.3.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
